Start work on rearranging things so that we can seperate the test code from the code needed to do the Test App UI
