+++ TestApp/TestApp.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\nac.Forms.UITesterApp\nac.Forms.UITesterApp.csproj" />
     <ProjectReference Include="..\nac.Forms\nac.Forms.csproj" />
   </ItemGroup>
