+++ nac.Forms.UITesterApp/model/LogViewerMessage.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TestApp.model;
+namespace nac.Forms.UITesterApp.model;
 
 internal class LogViewerMessage : nac.ViewModelBase.ViewModelBase
 {
