+++ nac.Forms.UITesterApp/nac.Forms.UITesterApp.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\nac.Forms\nac.Forms.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="nac.ViewModelBase" Version="1.0.1" />
+    </ItemGroup>
+
+</Project>
